+++ sdk/spring/spring-messaging-azure-eventhubs/src/main/java/com/azure/spring/messaging/eventhubs/implementation/support/converter/EventHubsMessageConverter.java
@@ -1,11 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.spring.messaging.eventhubs.support.converter;
+package com.azure.spring.messaging.eventhubs.implementation.support.converter;
 
 import com.azure.messaging.eventhubs.EventData;
 import com.azure.spring.messaging.AzureHeaders;
-import com.azure.spring.messaging.converter.AbstractAzureMessageConverter;
+import com.azure.spring.messaging.implementation.converter.AbstractJacksonAzureMessageConverter;
 import com.azure.spring.messaging.eventhubs.support.EventHubsHeaders;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.slf4j.Logger;
@@ -27,7 +27,7 @@ import static com.azure.spring.messaging.implementation.converter.ObjectMapperHo
  * A converter to turn a {@link Message} to {@link EventData} and vice versa.
  *
  */
-public class EventHubsMessageConverter extends AbstractAzureMessageConverter<EventData, EventData> {
+public class EventHubsMessageConverter extends AbstractJacksonAzureMessageConverter<EventData, EventData> {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(EventHubsMessageConverter.class);